xrCore/Threading: Fix Linux implementation of Event::Wait with timeout
Event::Wait(u32) should do the exact same thing than Window's
implementation, based on WaitForSingleObject. Instead of returning True
when the timer has run out, it should return False when (quoting
synchapi's documentation) "the time-out interval elapsed, and the
object's state is nonsignaled". This was the opposite of what it was
doing.