xrEngine: Fix concurrency issues when starting engine in OpenGL
While starting the engine, the main thread (let's call it A) and the
primary thread (let's call it B) have to switch back and forth the
ownership of the OpenGL context. For that, it uses some Events. The
problematic Event is deviceReadyToRun.

In theory, it should do this:
thread B waits deviceReadyToRun -> thread A sets deviceReadyToRun ->
thread B awakes and thread A waits deviceReadyToRun -> thread B sets
deviceReadToRun -> thread A awakes
But often it would do this:
thread B waits deviceReadyToRun -> thread A sets deviceReadyToRun -> A
waits deviceReadyToRun -> thread A awakes
So thread B would get blocked indefinitely.

To fix it, just introduce a new event, primaryReadyToRun.