Improve error handling in server's browser
If one browser can't connect to the master-list, it shouldn't fail the whole refreshing session - the servers from other browsers should be displayed in the serverlist normally. Additionally, after 'error while connecting to master-list' error the browser won't return it every time when we are trying to refresh it again.