Filter 'junk' servers from the server's list
Some servers don't respond to client's queries (for example because of firewall settings). Let's filter them because they are almost totally useless for us. We could never connect to such servers because the game version of the item is unknown and "always different" and (moreover) we have no idea about the real port where the server runs.