Set default SSAO mode enabled by default (closes #1739)
In COP there is inconsistency, because with the default settings we have
default SSAO enabled on first launch, but r2_ssao_mode doesn't match to
that – it is set to HDAO and displays that it's enabled, while it's not.
After exit, r2_ssao_mode makes HDAO enabled.

I also removed ssao_opt_data hardcoded on and offs, which make Clear Sky
crash. (HDAO in CS only can work with opt data enabled, but engine was
disabling it, making it crash 100%)

Might be related to #1454