Fixed UB after switch to C++20 (fixes #1785)
So, because of unity build "bool operator==(const CUITabButton* btn,
const shared_str& id)" function became visible in UITrackBar.cpp, where
we were comparing shared_str with nullptr. That nullptr was implicitly
converted to CUITabButton*, resulting in wrong function being called and
a crash, because of nullptr dereference in that operator== function.

Actually, shared_str should not be compared to nullptr, because it
should be considered as something like std::string, which can't be
compared to nullptr.