fix: `SMusicTrack::SetVolume` setting the final volume to either 0.0 or 1.0 (#1815)
This was caused by what probably is a typo as the final volume is calculated
as `volume * m_Volume` but then assigned to a bool (which C++ just allows you todo).

This was introduced in commit 673c6d81f1ed0798c26bd901ae8f3525aab6b50b.

Looking at the diff this is not the behaviour we had before.