Temporary switch back to C++17
I was asked to postpone the switch to C++20

This commit is somewhat ugly, because it enables C++17 for Linux and other platforms, while Windows still has C++20 enabled.
Why? Because I don't want to revert commits that added designated initializers usage. While GCC and Clang for Linux, etc. has support for them in C++17 as GNU language extension, MSVC supports them only in C++20 mode so we have to keep it enabled for Windows builds.
It's temporary after all, and CI will help us to identify if we are using C++20 too much or not.